requirements.txt
pandas
pyyaml
scikit-learn
mlflow
flask
-e .